% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-actiongroupbuttons.R
\name{actionGroupButtons}
\alias{actionGroupButtons}
\title{Actions Buttons Group Inputs}
\usage{
actionGroupButtons(inputIds, labels, status = "default", size = "normal",
  direction = "horizontal", fullwidth = FALSE)
}
\arguments{
\item{inputIds}{The \code{input}s slot that will be used to access the value, one for each button.}

\item{labels}{Display a text in the center of each button.}

\item{status}{Color of the buttons}

\item{size}{Size of the buttons ('xs', 'sm', 'normal', 'lg')}

\item{direction}{Horizontal or vertical}

\item{fullwidth}{If TRUE, fill the width of the parent div}
}
\value{
An actions buttons group control that can be added to a UI definition.
}
\description{
Create a group of actions buttons.
}
