% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw-dropdown.R
\name{animateOptions}
\alias{animateOptions}
\title{Animate options}
\usage{
animateOptions(enter = "fadeInDown", exit = "fadeOutUp", duration = 1)
}
\arguments{
\item{enter}{Animation name on appearance}

\item{exit}{Animation name on disappearance}

\item{duration}{Duration of the animation}
}
\value{
a list
}
\description{
Animate options
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {

dropdown(
 "Your contents goes here ! You can pass several elements",
 circle = TRUE, status = "danger", icon = icon("gear"), width = "300px",
 animate = animateOptions(enter = "fadeInDown", exit = "fadeOutUp", duration = 3)
)

}
}
\seealso{
\code{\link{animations}}
}
