% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw-dropdown.R
\docType{data}
\name{animations}
\alias{animations}
\title{Animation names}
\format{
A list of lists
}
\source{
\url{https://github.com/daneden/animate.css/blob/master/animate-config.json}
}
\usage{
animations
}
\description{
List of all animations by categories
}
\keyword{datasets}
