% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-autonumeric.R
\name{autonumericInput}
\alias{autonumericInput}
\title{Autonumeric Input Widget}
\usage{
autonumericInput(
  inputId,
  label,
  value,
  width = NULL,
  align = "center",
  currencySymbol = "",
  currencySymbolPlacement = "s",
  decimalCharacter = ".",
  digitGroupSeparator = ",",
  allowDecimalPadding = TRUE,
  decimalPlaces = 2,
  decimalPlacesRawValue = NULL,
  decimalPlacesShownOnBlur = NULL,
  decimalPlacesShownOnFocus = NULL,
  defaultValueOverride = NULL,
  digitalGroupSpacing = 3,
  alwaysAllowDecimalCharacter = FALSE,
  caretPositionOnFocus = "focus",
  createLocalList = TRUE,
  decimalCharacterAlternative = NULL,
  divisorWhenUnfocused = NULL,
  rawValueDivisor = NULL,
  emptyInputBehavior = "focus",
  selectNumberOnly = TRUE,
  selectOnFocus = TRUE,
  eventBubbles = TRUE,
  eventIsCancelable = TRUE,
  failOnUnknownOption = FALSE,
  formatOnPageLoad = TRUE,
  formulaMode = FALSE,
  historySize = 20,
  isCancellable = TRUE,
  leadingZero = "deny",
  maximumValue = 1e+12,
  minimumValue = -1e+12,
  modifyValueOnWheel = TRUE,
  wheelOn = "focus",
  wheelStep = "progressive",
  negativeBracketsTypeOnBlur = NULL,
  negativePositiveSignPlacement = NULL,
  negativeSignCharacter = "-",
  positiveSignCharacter = "+",
  showPositiveSign = FALSE,
  noEventListeners = FALSE,
  onInvalidPaste = "error",
  outputFormat = "string",
  overrideMinMaxLimits = "ceiling",
  readOnly = FALSE,
  roundingMethod = "S",
  saveValueToSessionStorage = FALSE,
  serializeSpaces = "+",
  showOnlyNumbersOnFocus = FALSE,
  showWarnings = TRUE,
  styleRules = NULL,
  suffixText = NULL,
  symbolWhenUnfocused = "",
  unformatOnHover = TRUE,
  unformatOnSubmit = FALSE,
  valuesToStrings = NULL,
  watchExternalChanges = FALSE
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value (unformatted).}

\item{width}{The width of the input box, eg. \code{"200px"} or
\code{"100\%"}.}

\item{align}{The alignment of the text inside the input box, one of
"center", "left", "right". Defaults to "center".}

\item{watchExternalChanges}{}
}
\value{

}
\description{
An `R` wrapper over the javascript `AutoNumeric` library, for
formatting numeric inputs in shiny applications.
}
