% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseSliderSkin.R
\name{chooseSliderSkin}
\alias{chooseSliderSkin}
\title{Theme selector for sliderInput}
\usage{
chooseSliderSkin(skin = c("Shiny", "Flat", "Modern", "Nice", "Simple",
  "HTML5", "Round", "Square"))
}
\arguments{
\item{skin}{The \code{skin} to apply. Choose among 5 different flavors, namely 'Shiny', 'Flat', 'Modern', 'Nice', 'Simple", 'HTML5', 'Round' and 'Square'.}
}
\description{
Customize the appearance of the original shiny's sliderInputs
}
\note{
It is not currently possible to apply multiple themes at the same time.
}
\examples{
\dontrun{

if (interactive()) {

library(shiny)
library(shinyWidgets)

ui <- fluidPage(

  # use the Modern design
  chooseSliderSkin("Modern"),
  sliderInput("obs", "Customized slider1:",
              min = 0, max = 100, value = 50
  ),
  sliderInput("obs2", "Customized slider1:",
              min = 0, max = 100, value = 50
  ),
  sliderInput("obs3", "Customized slider1:",
              min = 0, max = 100, value = 50
  ),
  plotOutput("distPlot")
)

server <- function(input, output) {

  output$distPlot <- renderPlot({
    hist(rnorm(input$obs))
  })
}

shinyApp(ui, server)

}


}
}
\seealso{
See \code{\link{setSliderColor}} to update the color of your sliderInput.
}
