% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-circlebutton.R
\name{circleButton}
\alias{circleButton}
\title{Circle Action button}
\usage{
circleButton(inputId, icon = NULL, status = "default", size = "default",
  ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{icon}{An icon to appear on the button.}

\item{status}{Color of the button.}

\item{size}{Size of the button : default, lg, sm, xs.}

\item{...}{Named attributes to be applied to the button.}
}
\description{
Create a rounded action button.
}
