% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetalert.R
\name{closeSweetAlert}
\alias{closeSweetAlert}
\title{Close Sweet Alert}
\usage{
closeSweetAlert(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{The \code{session} object passed to function given
to shinyServer.}
}
\description{
Close Sweet Alert
}
