% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-utils.R
\name{create_tree}
\alias{create_tree}
\title{Create choice structure for \code{\link[=treeInput]{treeInput()}}}
\usage{
create_tree(data, levels = names(data), levels_id = NULL, ...)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{levels}{Variables identifying hierarchical levels,
values of those variables will be used as text displayed.}

\item{levels_id}{Variable to use as ID for nodes. Careful! Spaces are not allowed in IDs.}

\item{...}{Addtional arguments.}
}
\value{
a \code{list} that can be used in \code{\link[=treeInput]{treeInput()}}.
}
\description{
Create choice structure for \code{\link[=treeInput]{treeInput()}}
}
