% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-autonumeric.R
\name{currencyInput}
\alias{currencyInput}
\title{Currency Input Widget}
\usage{
currencyInput(
  inputId,
  label,
  value,
  format = "euro",
  width = NULL,
  align = "center"
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value (unformatted).}

\item{format}{A character string specifying the currency format of the
input.  See "Details" for possible values.}

\item{width}{The width of the input box, eg. \code{"200px"} or
\code{"100\%"}.}

\item{align}{The alignment of the text inside the input box, one of
"center", "left", "right". Defaults to "center".}
}
\value{
a currency input widget that can be added to the UI of a shiny app.
}
\description{
A shiny widget for as-you-type formatting of currency. For a more modifiable
version see \code{\link{autonumericInput}}.
}
\details{

}
\examples{
if (interactive()) {
  library(shiny)
  library(shinyWidgets)

  ui <- fluidPage(
    tags$h2("Currency Input"),

    shinyWidgets:::currencyInput("id1", "Euro:", value = 1234, format = "euro", width = 200, align = "right"),
    verbatimTextOutput("res1"),

    shinyWidgets:::currencyInput("id2", "Dollar:", value = 1234, format = "dollar", width = 200, align = "left"),
    verbatimTextOutput("res2")
  )

  server <- function(input, output, session) {
    output$res1 <- renderPrint(input$id1)
    output$res2 <- renderPrint(input$id2)
  }

  shinyApp(ui, server)

}
}
\seealso{
\code{\link{autonumericInput}}
}
