% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-datepicker.R
\name{datepickerInput}
\alias{datepickerInput}
\title{Date picker}
\usage{
datepickerInput(
  inputId,
  label,
  value = NULL,
  range = FALSE,
  autoRanges = FALSE,
  autoApply = TRUE,
  startView = NULL,
  format = "YYYY-MM-DD",
  lang = "en-US",
  locale = NULL,
  firstDay = 1,
  grid = NULL,
  calendars = NULL,
  months = FALSE,
  years = FALSE,
  minYear = 1950,
  maxYear = NULL,
  minDate = NULL,
  maxDate = NULL,
  minDays = NULL,
  maxDays = NULL,
  filterDays = NULL,
  weekNumbers = FALSE,
  delimiter = " - ",
  tooltip = TRUE,
  tooltipText = c("day", "days"),
  rangesPosition = NULL,
  rangesLabels = NULL,
  zIndex = 10,
  readonly = TRUE,
  inline = FALSE,
  width = NULL,
  element = NULL
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value, a \code{Date} or a \code{character} in same format as \code{format} argument.
Use a vector of length two if \code{range = TRUE}.}

\item{range}{Activate range mode.}

\item{autoRanges}{Show predefined ranges to quickly select a range of dates.}

\item{autoApply}{Hide the apply and cancel buttons, and automatically
apply a new date range as soon as two dates are clicked.}

\item{startView}{Date shown in calendar when date picker is opened.}

\item{format}{The default output format. See
\href{https://easepick.com/packages/datetime.html#tokens-format}{online documentation}
for possible values.}

\item{lang}{Language. This option affect to day names and month names.}

\item{locale}{Icon and text for buttons. Named list where possible names are:
\code{nextMonth} / \code{previousMonth} (newt and previous buttons, default is chevron icon),
\code{cancel} / \code{apply} (text for buttons when \code{autoApply = FALSE}).}

\item{firstDay}{Day of start week. (0 - Sunday, 1 - Monday, 2 - Tuesday, etc...).}

\item{grid}{Number of calendar columns.}

\item{calendars}{Number of visible months.}

\item{months, years}{Enable dropdowns to select month or year.}

\item{minYear, maxYear}{Minimal and maximal year to select from when \code{years = TRUE}.}

\item{minDate, maxDate}{The minimum/earliest (resp. maximum/latest) date that can be selected.}

\item{minDays, maxDays}{The minimum (resp. maximum) days of the selected range.}

\item{filterDays}{A JavaScript function (as a character string) to mark days that cannot be selected.}

\item{weekNumbers}{Show week numbers.}

\item{delimiter}{Delimiter between dates (in range mode).}

\item{tooltip}{Showing tooltip with how much days will be selected (in range mode).}

\item{tooltipText}{Text for the tooltip.
First value is used for singular, second for plural: \code{c("day", "days")}.}

\item{rangesPosition}{Position of predefined ranges block.}

\item{rangesLabels}{Labels of predefined ranges buttons.
Default is \code{c("Today", "Yesterday", "Last 7 Days", "Last 30 Days", "This Month", "Last Month")}.}

\item{zIndex}{zIndex of picker.}

\item{readonly}{Add readonly attribute to input element.}

\item{inline}{Show calendar inline.}

\item{width}{Width of input field.}

\item{element}{Custom element as a \code{shiny.tag} (typically a \code{tags$button()}
for example) on which attach the calendar. The element will have an id set as \verb{<inputId>_trigger}.}
}
\value{
A \code{Date} server-side, default value is \code{NULL} unless \code{value} is provided.
}
\description{
Single or range date picker with predefined ranges included,
multiple calendars grid display and other features.
Made with the JavaScript library \href{https://easepick.com/}{easepick}.
}
\note{
See \href{https://easepick.com/packages/core.html#options}{JavaScript documentation}
for more examples and options. Note that not all plugins are integrated in this widget.
}
\seealso{
\code{\link[=updateDatepickerInput]{updateDatepickerInput()}} to update value server-side.
}
