% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{demoAirDatepicker}
\alias{demoAirDatepicker}
\title{Some examples on how to use airDatepickerInput}
\usage{
demoAirDatepicker(example = "datepicker")
}
\arguments{
\item{example}{Name of the example : \code{"datepicker"},
\code{"timepicker"}, \code{"months"}, \code{"years"},
\code{"update"}.}
}
\description{
Some examples on how to use airDatepickerInput
}
\examples{
if (interactive()) {

  demoAirDatepicker("datepicker")

}
}
