% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{demoNoUiSlider}
\alias{demoNoUiSlider}
\title{Some examples on how to use noUiSliderInput}
\usage{
demoNoUiSlider(example = "color")
}
\arguments{
\item{example}{Name of the example : \code{"color"},
\code{"update"}, \code{"behaviour"}, \code{"more"},
\code{"format"}.}
}
\description{
Some examples on how to use noUiSliderInput
}
\examples{
if (interactive()) {

  demoNoUiSlider("color")

}
}
