% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{demoNumericRange}
\alias{demoNumericRange}
\title{An example showing how numericRangeInput works}
\usage{
demoNumericRange()
}
\description{
An example showing how numericRangeInput works
}
\examples{
if (interactive()) {

  demoNumericRange()

}
}
