% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual-select.R
\name{demoVirtualSelect}
\alias{demoVirtualSelect}
\title{Demo for \code{\link[=virtualSelectInput]{virtualSelectInput()}}}
\usage{
demoVirtualSelect(
  name = c("default", "update", "choices-format", "prepare-choices")
)
}
\arguments{
\item{name}{Name of the demo app to launch.}
}
\value{
No value.
}
\description{
Demo for \code{\link[=virtualSelectInput]{virtualSelectInput()}}
}
\examples{
\dontrun{

# Default usage
demoVirtualSelect("default")

# Update widget from server
demoVirtualSelect("update")

# Differents ways of specifying choices
demoVirtualSelect("choices-format")

# Prepare choices from a data.frame
demoVirtualSelect("prepare-choices")

}
}
