% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{useShinydashboard}
\alias{useShinydashboardPlus}
\alias{useTablerDash}
\alias{useArgonDash}
\alias{useBs4Dash}
\alias{setShadow}
\alias{setSliderColor}
\title{Deprecated function}
\usage{
useShinydashboard()

useShinydashboardPlus()

useTablerDash()

useArgonDash()

useBs4Dash(...)

setShadow(id = NULL, class = NULL)

setSliderColor(color, sliderId)
}
\arguments{
\item{...}{Deprecated arguments}

\item{id, class}{Deprecated arguments}

\item{color, sliderId}{Deprecated arguments}
}
\description{
Those functions are deprecated and will be removed in a futur release of shinyWidgets
}
