% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-menu.R
\name{dropMenuOptions}
\alias{dropMenuOptions}
\title{Drop menu options}
\usage{
dropMenuOptions(duration = c(275, 250), animation = "fade", flip = FALSE, ...)
}
\arguments{
\item{duration}{Duration of the CSS transition animation in ms.}

\item{animation}{The type of transition animation.}

\item{flip}{Determines if the tippy flips so that it is placed within
the viewport as best it can be if there is not enough space.}

\item{...}{Additional arguements.}
}
\value{
a \code{list} of options to be used in \code{\link{dropMenu}}.
}
\description{
Those options will passed to the underlying JavaScript
 library powering \code{dropMenu} : tippy.js. See all available options
 here \url{https://atomiks.github.io/tippyjs/all-props/}.
}
