% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{html-dependencies}
\alias{html-dependencies}
\alias{html_dependency_awesome}
\alias{html_dependency_bttn}
\alias{html_dependency_pretty}
\alias{html_dependency_bsswitch}
\title{HTML dependencies}
\usage{
html_dependency_awesome()

html_dependency_bttn()

html_dependency_pretty()

html_dependency_bsswitch()
}
\value{
an \code{\link[htmltools]{htmlDependency}}.
}
\description{
These functions are used internally to load dependencies for widgets.
 Not all of them are exported. Below are the ones needed for package \href{https://github.com/dreamRs/fresh}{fresh}.
}
\examples{

# Use in UI or tags function

library(shiny)
fluidPage(
  html_dependency_awesome()
)

}
