% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-selectpicker.R
\name{pickerInputNEW}
\alias{pickerInputNEW}
\title{Select picker Input Control}
\usage{
pickerInputNEW(inputId, choices, selected = NULL, choicesOpt = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{choices}{List of values to select from. If elements of the
list are named then that name rather than the value is displayed to the user.}

\item{selected}{The initially selected value (or multiple values if multiple = TRUE).
If not specified then defaults to the first value for single-select lists
 and no values for multiple select lists.}

\item{choicesOpt}{Options for choices in the dropdown menu.}
}
\value{
A select control that can be added to a UI definition.
}
\description{
Create a select picker (\url{https://developer.snapappointments.com/bootstrap-select/})
}
