% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-background.R
\name{setBackgroundColor}
\alias{setBackgroundColor}
\title{Custom background color for your shinyapp}
\usage{
setBackgroundColor(color = "ghostwhite")
}
\arguments{
\item{color}{Background color. Use either the fullname or the Hex code
(\url{https://www.w3schools.com/colors/colors_hex.asp}).}
}
\description{
Allow to change the background color of your shinyapp.
}
\examples{
\dontrun{

if (interactive()) {

library(shiny)

ui <- fluidPage(
  tags$h2("Change shiny app background"),
  setBackgroundColor("ghostwhite")
)

server <- function(input, output, session) {

}

shinyApp(ui, server)

}

}
}
