% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyWidgetsGallery.R
\name{shinyWidgetsGallery}
\alias{shinyWidgetsGallery}
\title{Launch the shinyWidget Gallery}
\usage{
shinyWidgetsGallery()
}
\description{
A gallery of widgets available in the package.
}
\examples{
if (interactive()) {

 shinyWidgetsGallery()

}
}
