% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-dropdown.R
\name{toggleDropdownButton}
\alias{toggleDropdownButton}
\title{Toggle a dropdown menu}
\usage{
toggleDropdownButton(inputId)
}
\arguments{
\item{inputId}{Id for the dropdown to toggle}
}
\description{
Open or close a dropdown menu server-side
}
