% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-input.R
\name{treeInput}
\alias{treeInput}
\title{Tree Input Widget}
\usage{
treeInput(
  inputId,
  label,
  choices,
  selected = NULL,
  closeDepth = 1,
  returnValue = c("text", "id", "all"),
  width = NULL
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{choices}{A \code{list} of \code{list} in a tree structure, see \code{\link[=create_tree]{create_tree()}} for examples creating the right structure.}

\item{selected}{Inital selected values, note that you have to use node ID.}

\item{closeDepth}{Expand level, default to only first one visible.}

\item{width}{The width of the input, e.g. \verb{400px}, or \verb{"100\%}.}
}
\value{
A \code{shiny.tag} object that can be used in a UI definition.
}
\description{
A tree input widget allowing to select values in a hierarchical structure.
}
