% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-autonumeric.R
\name{updateAutonumericInput}
\alias{updateAutonumericInput}
\title{Update an Autonumeric Input Object}
\usage{
updateAutonumericInput(
  session,
  inputId,
  label = NULL,
  value = NULL,
  options = NULL
)
}
\arguments{
\item{session}{Standard shiny \code{session}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{The value to set for the input object.}

\item{options}{List of additional parameters to update, use
\code{autonumericInput}'s arguments.}
}
\description{
Update an Autonumeric Input Object
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyWidgets)

 ui <- fluidPage(
   h1("AutonumericInput Update Example"),
   br(),
   shinyWidgets:::autonumericInput(
     inputId = "id1",
     label = "Autonumeric Input",
     value = 1234.56,
     align = "center",
     currencySymbol = "$ ",
     currencySymbolPlacement = "p",
     decimalCharacter = ".",
     digitGroupSeparator = ",",
     divisorWhenUnfocused = 1000,
     symbolWhenUnfocused = "K",
     decimalPlacesShownOnBlur = 1
   ),
   verbatimTextOutput("res1"),
   actionButton("change", "Update autonumeric Input")
 )

 server <- function(input, output, session) {
   output$res1 <- renderPrint(input$id1)

   observeEvent(input$change, {
     updateAutonumericInput(
       session = session,
       inputId = "id1",
       label = "Updated Label!",
       value = 6543.21,
       options = list(
         currencySymbol = " #",
         currencySymbolPlacement = "s",
         decimalPlacesShownOnBlur = 2,
         symbolWhenUnfocused = NULL,
         divisorWhenUnfocused = NULL
       )
     )
   })
 }

 shinyApp(ui, server)
}
}
