% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-awesomecheckbox.R
\name{updateAwesomeCheckboxGroup}
\alias{updateAwesomeCheckboxGroup}
\title{Change the value of a radio input on the client}
\usage{
updateAwesomeCheckboxGroup(session, inputId, label = NULL, choices = NULL,
  selected = NULL, inline = FALSE, status = "primary")
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{label}{Input label.}

\item{choices}{List of values to show checkboxes for.}

\item{selected}{The values that should be initially selected, if any.}

\item{inline}{If TRUE, render the choices inline (i.e. horizontally)}

\item{status}{Color of the buttons}
}
\description{
Change the value of a radio input on the client
}
