% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar-pro-input.R
\name{updateCalendarPro}
\alias{updateCalendarPro}
\title{Update calendar pro from server}
\usage{
updateCalendarPro(
  inputId,
  label = NULL,
  value = NULL,
  mode = NULL,
  ...,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{mode}{This parameter determines whether selecting one or multiple days is allowed, or if date selection is completely disabled.
Possible values are: 'single' | 'multiple' | 'multiple-ranged' | false.}

\item{...}{Other settings passed to Calendar Pro JavaScript method, see \href{https://vanilla-calendar.pro/docs/reference/settings}{online documentation} for reference.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}
}
\value{
No value.
}
\description{
Update a \code{\link[=calendarProInput]{calendarProInput()}} from the server.
}
\examples{

library(shiny)
library(shinyWidgets)

ui <- fluidPage(
  theme = bslib::bs_theme(5),
  tags$h2("Calendar Pro Input: update from server"),
  fluidRow(
    column(
      width = 6,
      calendarProInput(
        inputId = "calendar",
        label = "Select a date:",
        placeholder = "Select a date",
        width = "100\%"
      ),
      verbatimTextOutput("res1"),
      textInput(
        inputId = "label",
        label = "Update label:"
      ),
      actionButton(
        inputId = "today",
        label = "Set value as today"
      ),
      actionButton(
        inputId = "today3",
        label = "Set value as today + 3"
      ),
      radioButtons(
        inputId = "mode",
        label = "Update mode:",
        choices = c("single", "multiple", "multiple-ranged"),
        inline = TRUE
      )
    ),
    column(
      width = 6
    )
  )
)

server <- function(input, output, session) {

  output$res1 <- renderPrint(input$calendar)

  observeEvent(input$label, {
    if (isTruthy(input$label)) {
      updateCalendarPro(inputId = "calendar", label = input$label)
    }
  })

  observeEvent(input$today, {
    updateCalendarPro(inputId = "calendar", value = Sys.Date())
  })

  observeEvent(input$today3, {
    updateCalendarPro(inputId = "calendar", value = Sys.Date() + 3)
  })

  observeEvent(input$mode, {
    updateCalendarPro(inputId = "calendar", selectionDatesMode = input$mode)
  }, ignoreInit = TRUE)

}

if (interactive())
  shinyApp(ui, server)
}
\seealso{
\code{\link[=calendarProInput]{calendarProInput()}} for creating a widget in the UI.
}
