% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checkboxgroupbuttons.R
\name{updateCheckboxGroupButtons}
\alias{updateCheckboxGroupButtons}
\title{Change the value of a checkboxes group buttons input on the client}
\usage{
updateCheckboxGroupButtons(session, inputId, selected = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{selected}{The values selected.}
}
\description{
Change the value of a radio group buttons input on the client
}
