% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-autonumeric.R
\name{updateCurrencyInput}
\alias{updateCurrencyInput}
\title{Update a Currency Input Widget}
\usage{
updateCurrencyInput(
  session,
  inputId,
  label = NULL,
  value = NULL,
  format = NULL
)
}
\arguments{
\item{session}{Standard shiny \code{session}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{The value to set for the input object.}

\item{format}{The format to change the input object to.}
}
\description{
Update a Currency Input Widget
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinyWidgets)

  ui <- fluidPage(
    tags$h2("Currency Input"),

    shinyWidgets:::currencyInput(
      inputId = "id1",
      label = "Euro:",
      value = 1234,
      format = "euro",
      width = 200,
      align = "right"
    ),
    verbatimTextOutput("res1"),
    actionButton("bttn1", "Change Euros to Dollars")
  )

  server <- function(input, output, session) {

    output$res1 <- renderPrint(input$id1)

    observeEvent(input$bttn1, {
      updateCurrencyInput(
        session = session,
        inputId = "id1",
        label = "Dollar:",
        format = "dollar"
      )
    })


  }

  shinyApp(ui, server)
}
}
