% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-datepicker.R
\name{updateDatepicker}
\alias{updateDatepicker}
\title{Change the value of \code{\link[=datepickerInput]{datepickerInput()}} on the client}
\usage{
updateDatepicker(
  inputId,
  label = NULL,
  value = NULL,
  range = NULL,
  minDate = NULL,
  maxDate = NULL,
  minDays = NULL,
  maxDays = NULL,
  filterDays = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{The id of the input object.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value, a \code{Date} or a \code{character} in same format as \code{format} argument.
Use a vector of length two if \code{range = TRUE}.}

\item{range}{Activate range mode.}

\item{minDate, maxDate}{The minimum/earliest (resp. maximum/latest) date that can be selected.}

\item{minDays, maxDays}{The minimum (resp. maximum) days of the selected range.}

\item{filterDays}{A JavaScript function (as a character string) to mark days that cannot be selected.}

\item{session}{The session object passed to function given to shinyServer.}
}
\value{
No value.
}
\description{
Change the value of \code{\link[=datepickerInput]{datepickerInput()}} on the client
}
