% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-knob.R
\name{updateKnobInput}
\alias{updateKnobInput}
\title{Change the value of a knob input on the client}
\usage{
updateKnobInput(session, inputId, label = NULL, value = NULL,
  readOnly = NULL)
}
\arguments{
\item{session}{Standard shiny \code{session}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{The value to set for the input object.}

\item{readOnly}{Don't work}
}
\description{
Change the value of a knob input on the client
}
\examples{
\dontrun{

if (interactive()) {

# TODO

}

}
}
