% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-materialswitch.R
\name{updateMaterialSwitch}
\alias{updateMaterialSwitch}
\title{Change the value of a materialSwitch input on the client}
\usage{
updateMaterialSwitch(session, inputId, value = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{value}{The value to set for the input object.}
}
\description{
Change the value of a materialSwitch input on the client
}
\seealso{
\code{\link{materialSwitch}}
}
