% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-noUiSlider.R
\name{updateNoUiSliderInput}
\alias{updateNoUiSliderInput}
\title{Change the value of a no ui slider input on the client}
\usage{
updateNoUiSliderInput(
  session = getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  range = NULL,
  disable = FALSE
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{value}{The new value.}

\item{range}{The new range, must be of length 2 with \code{c(min, max)}.}

\item{disable}{logical, disable or not the slider,
if disabled the user can no longer modify the slider value}
}
\description{
Change the value of a no ui slider input on the client
}
\examples{
if (interactive()) {

 demoNoUiSlider("update")

}
}
