% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-selectpicker.R
\name{updatePickerInput}
\alias{updatePickerInput}
\title{Change the value of a select picker input on the client}
\usage{
updatePickerInput(session, inputId, label = NULL, selected = NULL,
  choices = NULL, choicesOpt = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{label}{Display a text in the center of the switch.}

\item{selected}{The initially selected value (or multiple values if multiple = TRUE). If not specified then defaults to the first value for single-select lists and no values for multiple select lists.}

\item{choices}{List of values to select from. If elements of the list are named then that name rather than the value is displayed to the user.}

\item{choicesOpt}{Options for choices in the dropdown menu}
}
\description{
Change the value of a picker input on the client
}
