% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressBars.R
\name{updateProgressBar}
\alias{updateProgressBar}
\title{Update a progress bar}
\usage{
updateProgressBar(session, id, value, total = NULL, status = NULL)
}
\arguments{
\item{session}{The `session` object passed to function given to shinyServer.}

\item{id}{The id of the progress bar to update}

\item{value}{Value of the progress bar between 0 and 100, if >100 you must provide total}

\item{total}{Used to calculate percentage if value > 100}

\item{status}{Color}
}
\description{
Change the value of a progress bar on the client
}
