% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-radiogroupbuttons.R
\name{updateRadioGroupButtons}
\alias{updateRadioGroupButtons}
\title{Change the value of a radio group buttons input on the client}
\usage{
updateRadioGroupButtons(session, inputId, selected = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{selected}{The value selected.}
}
\description{
Change the value of a radio group buttons input on the client
}
