% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-switch.R
\name{updateSwitchInput}
\alias{updateSwitchInput}
\title{Change the value of a switch input on the client}
\usage{
updateSwitchInput(session, inputId, value = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{value}{TRUE or FALSE.}
}
\description{
Change the value of a switch input on the client
}
