% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-input.R
\name{updateTreeInput}
\alias{updateTreeInput}
\title{Update Tree Input}
\usage{
updateTreeInput(
  inputId,
  label = NULL,
  selected = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{selected}{The values that should be initially selected, if any.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}
}
\value{
No value.
}
\description{
Update \code{\link[=treeInput]{treeInput()}} from server.
}
