% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyWidgets.R
\docType{package}
\name{shinyWidgets}
\alias{shinyWidgets}
\alias{shinyWidgets-package}
\title{shinyWidgets: Custom inputs widgets for Shiny.}
\description{
The shinyWidgets package provides several custom widgets to extend those available in package shiny
}
\examples{
if (interactive()) {
  shinyWidgets::shinyWidgetsGallery()
}

}
