% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-dropdown.R
\name{tooltipOptions}
\alias{tooltipOptions}
\title{Tooltip options}
\usage{
tooltipOptions(placement = "right", title = "Params", html = FALSE)
}
\arguments{
\item{placement}{Placement of tooltip : right, top, bottom, left.}

\item{title}{Text of the tooltip}

\item{html}{Logical, allow HTML tags inside tooltip}
}
\description{
List of options for tooltip for a dropdown menu button.
}
