% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-airDatepicker.R
\name{updateAirDateInput}
\alias{updateAirDateInput}
\title{Change the value of \code{\link{airDatepickerInput}} on the client}
\usage{
updateAirDateInput(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  value = NULL,
  clear = FALSE,
  options = NULL,
  show = FALSE,
  hide = FALSE
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{The value to set for the input object.}

\item{clear}{Logical, clear all previous selected dates.}

\item{options}{Options to update, see available ones here: \url{http://t1m0n.name/air-datepicker/docs/}.}

\item{show, hide}{Show / hide datepicker.}
}
\description{
Change the value of \code{\link{airDatepickerInput}} on the client
}
\examples{
if (interactive()) {

  demoAirDatepicker("update")

}
}
