% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetalert.R
\name{useSweetAlert}
\alias{useSweetAlert}
\title{Load Sweet Alert dependencies}
\usage{
useSweetAlert()
}
\description{
This function isn't necessary for \code{sendSweetAlert}, \code{confirmSweetAlert},
 \code{inputSweetAlert}, but is still needed for \code{progressSweetAlert}.
}
\seealso{
\code{\link{sendSweetAlert}}, \code{\link{confirmSweetAlert}}, \code{\link{inputSweetAlert}}.
}
