% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetalert.R
\name{useSweetAlert}
\alias{useSweetAlert}
\title{Load Sweet Alert Dependancies}
\usage{
useSweetAlert()
}
\description{
This function is useless for \code{sendSweetAlert}, \code{confirmSweetAlert},
 \code{inputSweetAlert}, but is still needed for \code{progressSweetAlert}.
}
\note{
Use \code{receiveSweetAlert()} in the UI and \code{sendSweetAlert()} in the server.
}
\seealso{
\code{\link{sendSweetAlert}}, \code{\link{confirmSweetAlert}}, \code{\link{inputSweetAlert}}
}
